requirements.txt
flask
flask_restful
flask_sqlalchemy
marshmallow
flask_marshmallow  
marshmallow-sqlalchemy
